Абстрактный класс и наследование
