+++ src/waypoint-edit.js
@@ -1,44 +1,12 @@
-class WaypointEdit {
+import {WaypointComponent} from "./waypointComponent";
+
+class WaypointEdit extends WaypointComponent {
   constructor(data) {
-    this._type = data.type;
+    super(data);
     this._typeList = data.typeList;
-    this._destination = data.destination;
-    this._time = data.time;
-    this._duration = data.duration;
-    this._price = data.price;
-    this._photo = data.photo;
-    this._offers = data.offers;
-    this._text = data.text;
-    this._element = null;
 
     this._onSubmit = null;
     this._onReset = null;
-  }
-
-  /**
-   * Генерирует html-код оферов
-   * @param {array} offerList
-   * @return {array}
-   */
-  _generateOffersHtml(offerList) {
-    return offerList.reduce((resultHtml, offerItem) => {
-      return resultHtml + `
-      <li>
-        <button class="trip-point__offer">${offerItem}</button>
-      </li>
-    `;
-    }, ``);
-  }
-
-  /**
-   * Генерирует html - код текстовых предложений
-   * @param {array} textList
-   * @return {string}
-   */
-  _generateTextHtml(textList) {
-    return textList.reduce((resultHtml, textItem) => {
-      return resultHtml + `${textItem}<br>`;
-    }, ``);
   }
 
   /**
@@ -73,118 +41,105 @@ class WaypointEdit {
    */
   get template() {
     return `
-    <!--<article class="trip-point">
-      <i class="trip-icon">${this._type.icon}</i>
-      <h3 class="trip-point__title">${this._type.name} to ${this._destination}<br><br>${this._generateTextHtml(this._text)}</h3>
-      <p class="trip-point__schedule">
-        <span class="trip-point__timetable">${this._time}</span>
-        <span class="trip-point__duration">${this._duration}</span>
-      </p>
-      <p class="trip-point__price">&euro;&nbsp;${this._price}</p>
-      <ul class="trip-point__offers">
-      ${this._generateOffersHtml(this._offers)}
-      </ul>
-    </article>-->
+      <article class="point">
+        <form action="" method="get">
+          <header class="point__header">
+            <label class="point__date">
+              choose day
+              <input class="point__input" type="text" placeholder="MAR 18" name="day">
+            </label>
 
-<article class="point">
-  <form action="" method="get">
-    <header class="point__header">
-      <label class="point__date">
-        choose day
-        <input class="point__input" type="text" placeholder="MAR 18" name="day">
-      </label>
+            <div class="travel-way">
+              <label class="travel-way__label" for="travel-way__toggle">${this._type.icon}</label>
 
-      <div class="travel-way">
-        <label class="travel-way__label" for="travel-way__toggle">${this._type.icon}</label>
+              <input type="checkbox" class="travel-way__toggle visually-hidden" id="travel-way__toggle">
 
-        <input type="checkbox" class="travel-way__toggle visually-hidden" id="travel-way__toggle">
+              <div class="travel-way__select">
+                <div class="travel-way__select-group">
+                ${this._generateTypeListHtml(this._type, this._typeList)}
+                </div>
 
-        <div class="travel-way__select">
-          <div class="travel-way__select-group">
-          ${this._generateTypeListHtml(this._type, this._typeList)}
-          </div>
+                <div class="travel-way__select-group">
+                  <input class="travel-way__select-input visually-hidden" type="radio" id="travel-way-check-in" name="travel-way" value="check-in">
+                  <label class="travel-way__select-label" for="travel-way-check-in">🏨 check-in</label>
 
-          <div class="travel-way__select-group">
-            <input class="travel-way__select-input visually-hidden" type="radio" id="travel-way-check-in" name="travel-way" value="check-in">
-            <label class="travel-way__select-label" for="travel-way-check-in">🏨 check-in</label>
+                  <input class="travel-way__select-input visually-hidden" type="radio" id="travel-way-sightseeing" name="travel-way" value="sight-seeing">
+                  <label class="travel-way__select-label" for="travel-way-sightseeing">🏛 sightseeing</label>
+                </div>
+              </div>
+            </div>
 
-            <input class="travel-way__select-input visually-hidden" type="radio" id="travel-way-sightseeing" name="travel-way" value="sight-seeing">
-            <label class="travel-way__select-label" for="travel-way-sightseeing">🏛 sightseeing</label>
-          </div>
-        </div>
-      </div>
+            <div class="point__destination-wrap">
+              <label class="point__destination-label" for="destination">${this._type.name} to</label>
+              <input class="point__destination-input" list="destination-select" id="destination" value="${this._destination}" name="destination">
+              <datalist id="destination-select">
+                <option value="airport"></option>
+                <option value="Geneva"></option>
+                <option value="Chamonix"></option>
+                <option value="hotel"></option>
+              </datalist>
+            </div>
 
-      <div class="point__destination-wrap">
-        <label class="point__destination-label" for="destination">${this._type.name} to</label>
-        <input class="point__destination-input" list="destination-select" id="destination" value="${this._destination}" name="destination">
-        <datalist id="destination-select">
-          <option value="airport"></option>
-          <option value="Geneva"></option>
-          <option value="Chamonix"></option>
-          <option value="hotel"></option>
-        </datalist>
-      </div>
+            <label class="point__time">
+              choose time
+              <input class="point__input" type="text" value="${this._time}" name="time" placeholder="00:00 — 00:00">
+            </label>
 
-      <label class="point__time">
-        choose time
-        <input class="point__input" type="text" value="${this._time}" name="time" placeholder="00:00 — 00:00">
-      </label>
+            <label class="point__price">
+              write price
+              <span class="point__price-currency">€</span>
+              <input class="point__input" type="text" value="${this._price}" name="price">
+            </label>
 
-      <label class="point__price">
-        write price
-        <span class="point__price-currency">€</span>
-        <input class="point__input" type="text" value="${this._price}" name="price">
-      </label>
+            <div class="point__buttons">
+              <button class="point__button point__button--save" type="submit">Save</button>
+              <button class="point__button" type="reset">Delete</button>
+            </div>
 
-      <div class="point__buttons">
-        <button class="point__button point__button--save" type="submit">Save</button>
-        <button class="point__button" type="reset">Delete</button>
-      </div>
+            <div class="paint__favorite-wrap">
+              <input type="checkbox" class="point__favorite-input visually-hidden" id="favorite" name="favorite">
+              <label class="point__favorite" for="favorite">favorite</label>
+            </div>
+          </header>
 
-      <div class="paint__favorite-wrap">
-        <input type="checkbox" class="point__favorite-input visually-hidden" id="favorite" name="favorite">
-        <label class="point__favorite" for="favorite">favorite</label>
-      </div>
-    </header>
+          <section class="point__details">
+            <section class="point__offers">
+              <h3 class="point__details-title">offers</h3>
 
-    <section class="point__details">
-      <section class="point__offers">
-        <h3 class="point__details-title">offers</h3>
+              <div class="point__offers-wrap">
+                <input class="point__offers-input visually-hidden" type="checkbox" id="add-luggage" name="offer" value="add-luggage">
+                <label for="add-luggage" class="point__offers-label">
+                  <span class="point__offer-service">Add luggage</span> + €<span class="point__offer-price">30</span>
+                </label>
 
-        <div class="point__offers-wrap">
-          <input class="point__offers-input visually-hidden" type="checkbox" id="add-luggage" name="offer" value="add-luggage">
-          <label for="add-luggage" class="point__offers-label">
-            <span class="point__offer-service">Add luggage</span> + €<span class="point__offer-price">30</span>
-          </label>
+                <input class="point__offers-input visually-hidden" type="checkbox" id="switch-to-comfort-class" name="offer" value="switch-to-comfort-class">
+                <label for="switch-to-comfort-class" class="point__offers-label">
+                  <span class="point__offer-service">Switch to comfort class</span> + €<span class="point__offer-price">100</span>
+                </label>
 
-          <input class="point__offers-input visually-hidden" type="checkbox" id="switch-to-comfort-class" name="offer" value="switch-to-comfort-class">
-          <label for="switch-to-comfort-class" class="point__offers-label">
-            <span class="point__offer-service">Switch to comfort class</span> + €<span class="point__offer-price">100</span>
-          </label>
+                <input class="point__offers-input visually-hidden" type="checkbox" id="add-meal" name="offer" value="add-meal">
+                <label for="add-meal" class="point__offers-label">
+                  <span class="point__offer-service">Add meal </span> + €<span class="point__offer-price">15</span>
+                </label>
 
-          <input class="point__offers-input visually-hidden" type="checkbox" id="add-meal" name="offer" value="add-meal">
-          <label for="add-meal" class="point__offers-label">
-            <span class="point__offer-service">Add meal </span> + €<span class="point__offer-price">15</span>
-          </label>
+                <input class="point__offers-input visually-hidden" type="checkbox" id="choose-seats" name="offer" value="choose-seats">
+                <label for="choose-seats" class="point__offers-label">
+                  <span class="point__offer-service">Choose seats</span> + €<span class="point__offer-price">5</span>
+                </label>
+              </div>
 
-          <input class="point__offers-input visually-hidden" type="checkbox" id="choose-seats" name="offer" value="choose-seats">
-          <label for="choose-seats" class="point__offers-label">
-            <span class="point__offer-service">Choose seats</span> + €<span class="point__offer-price">5</span>
-          </label>
-        </div>
-
-      </section>
-      <section class="point__destination">
-        <h3 class="point__details-title">Destination</h3>
-        <p class="point__destination-text">Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.</p>
-        <div class="point__destination-images">
-          ${this._generatePhotoHtml(this._photo)}
-        </div>
-      </section>
-      <input type="hidden" class="point__total-price" name="total-price" value="">
-    </section>
-  </form>
-</article>
+            </section>
+            <section class="point__destination">
+              <h3 class="point__details-title">Destination</h3>
+              <p class="point__destination-text">Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.</p>
+              <div class="point__destination-images">
+                ${this._generatePhotoHtml(this._photo)}
+              </div>
+            </section>
+            <input type="hidden" class="point__total-price" name="total-price" value="">
+          </section>
+        </form>
+      </article>
   `;
   }
 
@@ -198,9 +153,6 @@ class WaypointEdit {
     if (typeof this._onReset === `function`) {
       this._onReset();
     }
-  }
-  get element() {
-    return this._element;
   }
 
   set onSubmit(fn) {
@@ -226,25 +178,6 @@ class WaypointEdit {
     const pointBtnResetElem = this._element.querySelector(`[type="reset"]`);
     pointBtnSubmitElem.removeEventListener(`click`, this._onWaypointSubmit.bind(this));
     pointBtnResetElem.removeEventListener(`click`, this._onWaypointReset.bind(this));
-  }
-
-  /**
-   * Отрисовка задачи в заданном элементе
-   * @return {link}
-   */
-  render() {
-    this._element = null || document.createElement(`div`);
-
-    this._element.innerHTML = this.template;
-    this._element = this._element.firstElementChild;
-    this.addEvents();
-
-    return this._element;
-  }
-
-  unrender() {
-    this.removeEvents();
-    this._element = null;
   }
 }
