+++ src/waypoint.js
@@ -1,42 +1,10 @@
-class Waypoint {
+import {WaypointComponent} from "./waypointComponent";
+
+class Waypoint extends WaypointComponent {
   constructor(data) {
-    this._type = data.type;
-    this._destination = data.destination;
-    this._time = data.time;
-    this._duration = data.duration;
-    this._price = data.price;
-    this._photo = data.photo;
-    this._offers = data.offers;
-    this._text = data.text;
-    this._element = null;
+    super(data);
 
     this._onClick = null;
-  }
-
-  /**
-   * Генерирует html-код оферов
-   * @param {array} offerList
-   * @return {array}
-   */
-  _generateOffersHtml(offerList) {
-    return offerList.reduce((resultHtml, offerItem) => {
-      return resultHtml + `
-      <li>
-        <button class="trip-point__offer">${offerItem}</button>
-      </li>
-    `;
-    }, ``);
-  }
-
-  /**
-   * Генерирует html - код текстовых предложений
-   * @param {array} textList
-   * @return {string}
-   */
-  _generateTextHtml(textList) {
-    return textList.reduce((resultHtml, textItem) => {
-      return resultHtml + `${textItem}<br>`;
-    }, ``);
   }
 
   /**
@@ -66,10 +34,6 @@ class Waypoint {
     }
   }
 
-  get element() {
-    return this._element;
-  }
-
   set onClick(fn) {
     this._onClick = fn;
   }
@@ -83,26 +47,6 @@ class Waypoint {
 
   removeEvents() {
     this._element.removeEventListener(`click`, this._onWaypointClick.bind(this));
-  }
-
-  /**
-   * Отрисовка задачи в заданном элементе
-   * @return {link}
-   */
-  render() {
-    this._element = null || document.createElement(`div`);
-
-    this._element.innerHTML = this.template;
-    this._element = this._element.firstElementChild;
-
-    this.addEvents();
-
-    return this._element;
-  }
-
-  unrender() {
-    this.removeEvents();
-    this._element = null;
   }
 }
